requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
python-whois-extended==0.7.12
requests==2.19.1
urllib3==1.23
